# docs/api/server.openapi.yaml
openapi: 3.0.3
info:
  title: 通用企业知识库平台 — web ↔ server API（v0.2 草稿）
  version: 0.2.0-draft
  description: |
    机器可读权威契约。**状态：评审中，未冻结**。本文件取代 v0.1 版本并覆盖 v0.2 设计能力；
    尚未覆盖的字段以 `TBD(评审确认)` 标注，不得据此生成生产代码。
    设计依据：docs/02-概要设计.md、docs/03-详细设计.md §16、docs/modules/enterprise-generalization/requirements。
servers:
  - url: https://api.ragkb.example.com
security:
  - oidcSession: []      # OIDC 生产：BFF 设置的 HttpOnly 会话 cookie（现状不动）
  - bearerJwt: []        # form 开发：JWT access token（Authorization: Bearer <token>）
...
components:
  securitySchemes:
    oidcSession:
      type: apiKey
      in: cookie
      name: ragkb_session
      description: BFF 设置的 Secure/HttpOnly/SameSite 会话 cookie；浏览器不保存长期 token（02-概要 §4.1）。
    bearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        form 模式的 JWT access token；前端仅内存持有（不落 localStorage），随请求头
        `Authorization: Bearer <token>` 发送。刷新凭证走 HttpOnly cookie `ragkb_refresh`
        （SameSite=Lax, Path=/api/v1/auth，生产加 Secure），不进入请求头；刷新见
        `POST /api/v1/auth/refresh`。
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: ApiKey
      description: '机器访问；格式 `Authorization: Bearer <apikey>`，scope 限制于 key 本身。'
    oidc:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://idp.example.com/authorize
          tokenUrl: https://idp.example.com/token
          scopes:
            openid: OpenID Connect
            profile: 用户基础资料
            email: 邮箱
  parameters:
    page:
      name: page
      in: query
      description: 页号（从 1 开始）
      schema: { type: integer, minimum: 1, default: 1 }
    size:
      name: size
      in: query
      description: 每页条数
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    idempotencyKey:
      name: Idempotency-Key
      in: header
      description: 幂等键，写操作必填；同一键 + 相同请求摘要返回原结果，不同摘要返回 409
      schema: { type: string, maxLength: 128 }
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message, data]
      properties:
        code: { type: string, example: "0" }
        message: { type: string }
        data: {}
    Error:
      type: object
      required: [code, message, requestId, timestamp]
      properties:
        code: { type: string, example: "E-1003" }
        message: { type: string }
        requestId: { type: string }
        timestamp: { type: string, format: date-time }
        details: { type: object, nullable: true }
    PageData:
      type: object
      properties:
        items: { type: array, items: {} }
        total: { type: integer }
        page: { type: integer }
        size: { type: integer }
        hasMore: { type: boolean }
...
    ApiKey:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        prefix: { type: string, description: "仅返回前缀，明文只出现一次" }
        scopes: { type: array, items: { type: string } }
        allowedKbIds: { type: array, items: { type: integer }, nullable: true, description: "null 表示不限 KB" }
        expiresAt: { type: string, format: date-time, nullable: true }
        lastUsedAt: { type: string, format: date-time, nullable: true }
        revokedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    ApiKeyCreateRequest:
      type: object
      required: [name, scopes]
      properties:
        name: { type: string, maxLength: 128 }
        scopes: { type: array, items: { type: string }, minItems: 1 }
        allowedKbIds: { type: array, items: { type: integer }, nullable: true }
        expiresAt: { type: string, format: date-time, nullable: true }
    ApiKeyCreated:
      allOf:
        - { $ref: "#/components/schemas/ApiKey" }
        - type: object
          properties:
            plaintext: { type: string, description: "明文 Key，仅此一次返回；服务端只存 HMAC digest" }
...
    WebhookSubscription:
      type: object
      properties:
        id: { type: integer, format: int64 }
        url: { type: string }
        events: { type: array, items: { type: string } }
        status: { type: string, enum: [ACTIVE, DISABLED] }
        createdAt: { type: string, format: date-time }
    WebhookSubscriptionRequest:
      type: object
      required: [url, events]
      properties:
        url: { type: string, description: "必须 https；经 SSRF 防护" }
        events: { type: array, items: { type: string }, minItems: 1 }
        secretRef: { type: string, description: "签名密钥引用，禁止明文" }
...
    Org:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        parentId: { type: integer, format: int64, nullable: true }
        memberCount: { type: integer }
    OrgCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 128 }
        parentId: { type: integer, format: int64, nullable: true }
    User:
      type: object
      description: 租户成员视图（管理中心；与全局 sys_user 分离展示）
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        email: { type: string }
        status: { type: string, enum: [ACTIVE, DISABLED] }
        role: { type: string, enum: [TENANT_ADMIN, SECURITY_ADMIN, KNOWLEDGE_ADMIN, AUDITOR, MEMBER] }
        orgName: { type: string, nullable: true }
        lastLoginAt: { type: string, format: date-time, nullable: true }
...
  responses:
...
    Forbidden:
      description: 无权限/被排除（403；防止枚举时返回 404）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
paths:
...
  /api/v1/api-keys:
    get:
      tags: [api-keys]
      summary: 列表（不含明文）
      security:
        - oidcSession: []
      responses:
        "200":
          description: API Key 列表
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/ApiKey" }
    post:
      tags: [api-keys]
      summary: 创建 scoped API Key（明文仅此一次返回）
      security:
        - oidcSession: []
      parameters:
        - { $ref: "#/components/parameters/idempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApiKeyCreateRequest" }
      responses:
        "201":
          description: 创建成功，返回含明文
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/ApiKeyCreated" }
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/api-keys/{keyId}:
    parameters:
      - name: keyId
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      tags: [api-keys]
      summary: 详情
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/ApiKey" }
    delete:
      tags: [api-keys]
      summary: 吊销（立即失效并清理授权缓存）
      responses:
        "204":
          description: 已吊销
  /api/v1/api-keys/{keyId}/rotate:
    post:
      tags: [api-keys]
      summary: 轮换：旧 key 短时保留，返回新明文
      parameters:
        - { $ref: "#/components/parameters/idempotencyKey" }
      responses:
        "201":
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/ApiKeyCreated" }
...
  /api/v1/audit-logs:
    get:
      tags: [audit]
      summary: 审计日志（AUDITOR/TENANT_ADMIN；追加写只读）
      parameters:
        - { $ref: "#/components/parameters/page" }
        - { $ref: "#/components/parameters/size" }
        - name: action
          in: query
          schema: { type: string }
        - name: resourceType
          in: query
          schema: { type: string }
        - name: actorId
          in: query
          schema: { type: integer, format: int64 }
        - name: dateFrom
          in: query
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/PageData" }
  /api/v1/webhook-subscriptions:
    get:
      tags: [integration]
      summary: Webhook 订阅列表
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/WebhookSubscription" }
    post:
      tags: [integration]
      summary: 创建订阅（URL 必须 https，经 SSRF 防护）
      parameters:
        - { $ref: "#/components/parameters/idempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WebhookSubscriptionRequest" }
      responses:
        "201":
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/WebhookSubscription" }
  /api/v1/webhook-subscriptions/{subscriptionId}:
    parameters:
      - name: subscriptionId
        in: path
        required: true
        schema: { type: integer, format: int64 }
    delete:
      tags: [integration]
      summary: 删除订阅
      responses:
        "204":
          description: 已删除
...
  /api/v1/orgs:
    get:
      tags: [integration]
      summary: 组织树（部门）
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Org" }
    post:
      tags: [integration]
      summary: 创建组织节点
      parameters:
        - { $ref: "#/components/parameters/idempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrgCreateRequest" }
      responses:
        "201":
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Org" }
  /api/v1/orgs/{orgId}:
    parameters:
      - name: orgId
        in: path
        required: true
        schema: { type: integer, format: int64 }
    patch:
      tags: [integration]
      summary: 更新组织（改名/移动父节点）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 128 }
                parentId: { type: integer, format: int64, nullable: true }
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Org" }
    delete:
      tags: [integration]
      summary: 删除组织（需先清空成员或迁移，触发策略版本增长）
      responses:
        "204":
          description: 已删除
...
  /api/v1/users:
    get:
      tags: [integration]
      summary: 租户成员列表（TENANT_ADMIN，分页）
      parameters:
        - { $ref: "#/components/parameters/page" }
        - { $ref: "#/components/parameters/size" }
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data:
                        allOf:
                          - { $ref: "#/components/schemas/PageData" }
                          - type: object
                            properties:
                              items:
                                type: array
                                items: { $ref: "#/components/schemas/User" }
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/users/{userId}/enable:
    parameters:
      - name: userId
        in: path
        required: true
        schema: { type: integer, format: int64 }
    post:
      tags: [integration]
      summary: 重新启用成员
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/User" }
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/users/{userId}/disable:
    parameters:
      - name: userId
        in: path
        required: true
        schema: { type: integer, format: int64 }
    post:
      tags: [integration]
      summary: 停用成员（撤权并清理授权缓存）
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/User" }
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/users/{userId}/org:
    parameters:
      - name: userId
        in: path
        required: true
        schema: { type: integer, format: int64 }
    patch:
      tags: [integration]
      summary: 调整成员所属组织
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orgId]
              properties:
                orgId: { type: integer, format: int64, nullable: true }
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/User" }
...
  /api/v1/notifications/{notificationId}/read:
    post:
      tags: [integration]
      summary: 标记单条通知已读
      parameters:
        - name: notificationId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "204":
          description: 已标记
...